Switched folder structure and built initial calc
The calculator types have been built out, and
it doesn't really function properly. The basic approach
is correct, but digits are causing issues.
Need to rethink how to display in particular.
I feel the tokenised inputs should be in a stack,
and the display generated from them each time the
tokenisation ends.
Then with the digits, rebuild into a number each time.
When a new digit is added, peek at the top of the
stack: if a number, pop that, concat and push.